zram shift, adreno include, audio read32 compat, mac80211 duplicate
- zram: cast prio to unsigned long before shifting by ZRAM_COMP_PRIORITY_BIT1
  (bit 32), same overflow as the mask.
- adreno_a6xx_gmu: set_dma_ops() needs <linux/dma-map-ops.h>.
- include/sound/soc-component.h: 5.10 renamed snd_soc_component_read32() to
  snd_soc_component_read(); add a thin inline under the old name rather than
  editing the 27 downstream codec files that use it.
- net/mac80211/agg-tx.c: ieee80211_tx_ba_session_handle_start() was defined
  twice; file now matches ACK.

+++ drivers/block/zram/zram_drv.c
@@ -167,7 +167,7 @@ static inline void zram_set_priority(struct zram *zram, u32 index, u32 prio)
 	 */
 	zram->table[index].flags &= ~(ZRAM_COMP_PRIORITY_MASK <<
 				      ZRAM_COMP_PRIORITY_BIT1);
-	zram->table[index].flags |= (prio << ZRAM_COMP_PRIORITY_BIT1);
+	zram->table[index].flags |= ((unsigned long)prio << ZRAM_COMP_PRIORITY_BIT1);
 }
 
 static inline u32 zram_get_priority(struct zram *zram, u32 index)
